Add support for polymorphic belongs_to for ERDs
MSP-9605

`reflection.klass` does not work on polymorphic belongs_to associations
because it attempts to use the name of the association as the class name
instead of returning all classes that work with the name.  So, to
calculate the classes, scan `ActiveRecord::Base.descedants` for any
polymorphic has_many which have the matching :as.