Make net-ssh a soft dependency
MSP-9881

`metasploit-credential` cannot declare `net-ssh` as a runtime dependency
because `metasploit-framework` includes its own version of `'net/ssh'`,
which would conflict with the `net-ssh` gem.  Therefore, move the
dependency on `net-ssh` to the Gemfile so the gem is only used for
`metasploit-credential`'s specs and `metasploit-framework`'s `'net/ssh'`
will be used in `metasploit-framework` and Metasploit Pro.